Send and accept friend requests based on email, view friend collections, chat with friends.

+++ lib/screens/plant/widgets/plant_photo.dart
@@ -8,7 +8,6 @@ import 'package:plant_collector/models/app_data.dart';
 import 'package:plant_collector/screens/plant/image.dart';
 import 'package:plant_collector/models/cloud_db.dart';
 import 'package:plant_collector/models/cloud_store.dart';
-import 'package:plant_collector/models/user.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:plant_collector/formats/colors.dart';
 
@@ -19,9 +18,13 @@ import 'package:plant_collector/formats/colors.dart';
 //grid icon will set image as thumbnail
 
 class PlantPhoto extends StatelessWidget {
+  final bool connectionLibrary;
   final String imageURL;
   final String imageDate;
-  PlantPhoto({@required this.imageURL, @required this.imageDate});
+  PlantPhoto(
+      {@required this.connectionLibrary,
+      @required this.imageURL,
+      @required this.imageDate});
 
   @override
   Widget build(BuildContext context) {
@@ -29,35 +32,52 @@ class PlantPhoto extends StatelessWidget {
       padding: EdgeInsets.all(4.0),
       child: GestureDetector(
         onLongPress: () {
-          showDialog(
-            context: context,
-            builder: (BuildContext context) {
-              return DialogConfirm(
-                title: 'Delete Image',
-                text: 'Are you sure you want to delete this image?',
-                onPressed: () async {
-                  //remove from database array
-                  await Provider.of<CloudDB>(context)
-                      .updateArrayInDocumentInCollection(
-                    arrayKey: kPlantImageList,
-                    action: false,
-                    entries: [imageURL],
-                    folder: kUserPlants,
-                    documentName:
-                        Provider.of<AppData>(context).forwardingPlantID,
-                  );
-                  //get reference from the provided URL
-                  StorageReference reference =
-                      await Provider.of<CloudStore>(context)
-                          .getReferenceFromURL(imageURL: imageURL);
-                  //delete image
-                  Provider.of<CloudStore>(context)
-                      .deleteImage(imageReference: reference);
-                  Navigator.pop(context);
-                },
-              );
-            },
-          );
+          if (connectionLibrary == false) {
+            showDialog(
+              context: context,
+              builder: (BuildContext context) {
+                return DialogConfirm(
+                  title: 'Delete Image',
+                  text: 'Are you sure you want to delete this image?',
+                  onPressed: () async {
+                    //*****REMOVE IMAGE*****//
+                    //remove from database array
+                    await Provider.of<CloudDB>(context)
+                        .updateArrayInDocumentInCollection(
+                      arrayKey: kPlantImageList,
+                      action: false,
+                      entries: [imageURL],
+                      folder: kUserPlants,
+                      documentName:
+                          Provider.of<AppData>(context).forwardingPlantID,
+                    );
+                    //get reference from the provided URL
+                    StorageReference reference =
+                        await Provider.of<CloudStore>(context)
+                            .getReferenceFromURL(imageURL: imageURL);
+                    //delete image
+                    Provider.of<CloudStore>(context)
+                        .deleteImage(imageReference: reference);
+                    //*****REMOVE THUMBNAIL*****//
+                    //get the thumbnail image name from the full sized image url
+                    String imageName = Provider.of<CloudStore>(context)
+                        .getThumbName(imageUrl: imageURL);
+                    //get the thumb ref
+                    StorageReference thumbRef = Provider.of<CloudStore>(context)
+                        .getImageRef(
+                            imageName: imageName,
+                            imageExtension: 'jpg',
+                            plantIDFolder: Provider.of<AppData>(context)
+                                .forwardingPlantID);
+                    //delete thumbnail
+                    Provider.of<CloudStore>(context)
+                        .deleteImage(imageReference: thumbRef);
+                    Navigator.pop(context);
+                  },
+                );
+              },
+            );
+          }
         },
         child: Container(
           margin: EdgeInsets.only(bottom: 10.0),
@@ -96,12 +116,12 @@ class PlantPhoto extends StatelessWidget {
                       Padding(
                         padding: EdgeInsets.all(5.0 *
                             MediaQuery.of(context).size.width *
-                            kTextScale),
+                            kScaleFactor),
                         child: Container(
                           color: Color(0x33000000),
                           padding: EdgeInsets.all(3 *
                               MediaQuery.of(context).size.width *
-                              kTextScale),
+                              kScaleFactor),
                           child: Text(
                             imageDate,
                             textAlign: TextAlign.center,
@@ -114,57 +134,54 @@ class PlantPhoto extends StatelessWidget {
                         ),
                       ),
                       Container(
-                        width:
-                            35 * MediaQuery.of(context).size.width * kTextScale,
-                        child: FlatButton(
-                          onPressed: () {
-                            showDialog(
-                              context: context,
-                              builder: (BuildContext context) {
-                                return DialogConfirm(
-                                  title: 'Plant Thumbnail',
-                                  text:
-                                      'Would you like to use this image as the plant thumbail image?',
-                                  onPressed: () async {
-                                    //ensure logged in user id is saved for path creation
-                                    Provider.of<CloudStore>(context)
-                                        .setUserFolderID(
-                                            (await Provider.of<UserAuth>(
-                                                        context)
-                                                    .getCurrentUser())
-                                                .uid);
-                                    //function bundle
-                                    String url = await Provider.of<CloudStore>(
-                                            context)
-                                        .thumbnailPackage(
-                                            imageURL: imageURL,
-                                            plantID:
-                                                Provider.of<AppData>(context)
-                                                    .forwardingPlantID);
-                                    //update the plant with new thumbnail url
-                                    print(Provider.of<AppData>(context)
-                                        .forwardingPlantID);
-                                    Provider.of<CloudDB>(context)
-                                        .updateDocumentInCollection(
-                                            data: Provider.of<CloudDB>(context)
-                                                .updatePairFull(
-                                                    key: kPlantThumbnail,
-                                                    value: url),
-                                            collection: kUserPlants,
-                                            documentName:
-                                                Provider.of<AppData>(context)
-                                                    .forwardingPlantID);
-                                    Navigator.pop(context);
-                                  },
-                                );
-                              },
-                            );
-                          },
-                          child: Icon(Icons.grid_on,
-                              color: Color(0x55FFFFFF),
-                              size: AppTextSize.medium *
-                                  MediaQuery.of(context).size.width),
-                        ),
+                        width: 38 *
+                            MediaQuery.of(context).size.width *
+                            kScaleFactor,
+                        child: connectionLibrary == false
+                            ? FlatButton(
+                                onPressed: () {
+                                  showDialog(
+                                    context: context,
+                                    builder: (BuildContext context) {
+                                      return DialogConfirm(
+                                        title: 'Plant Thumbnail',
+                                        text:
+                                            'Would you like to use this image as the plant thumbail image?',
+                                        onPressed: () async {
+                                          //run thumbnail package to get thumb url
+                                          String thumbUrl = await Provider.of<
+                                                  CloudStore>(context)
+                                              .thumbnailPackage(
+                                                  imageURL: imageURL,
+                                                  plantID: Provider.of<AppData>(
+                                                          context)
+                                                      .forwardingPlantID);
+                                          //set thumb url
+                                          Provider.of<CloudDB>(context)
+                                              .updateDocumentInCollection(
+                                                  data: Provider.of<CloudDB>(
+                                                          context)
+                                                      .updatePairFull(
+                                                          key: kPlantThumbnail,
+                                                          value: thumbUrl),
+                                                  collection: kUserPlants,
+                                                  documentName:
+                                                      Provider.of<AppData>(
+                                                              context)
+                                                          .forwardingPlantID);
+
+                                          Navigator.pop(context);
+                                        },
+                                      );
+                                    },
+                                  );
+                                },
+                                child: Icon(Icons.grid_on,
+                                    color: Color(0x55FFFFFF),
+                                    size: AppTextSize.huge *
+                                        MediaQuery.of(context).size.width),
+                              )
+                            : SizedBox(),
                       ),
                     ],
                   ),
